Move the testing to use a free function
