--- Sources/GenythingTest/Helpers/Result+isSuccess.swift
@@ -1,10 +1,0 @@
-import Foundation
-
-extension Result {
-    var isSuccess: Bool {
-        switch self {
-            case .success: return true
-            case .failure: return false
-        }
-    }
-}
