+++ Tests/GenythingTestTests/TestAllTests.swift
@@ -6,7 +6,7 @@ internal final class TestAllTests: XCTestCase {
 
     func test_all_passes_when_it_should_with_correct_iterations() {
         var count = 0
-        TestSuite().testAll(genTrue) { assert, value in
+        testAll(genTrue) { assert, value in
             count += 1
             assert(value)
         }
@@ -16,7 +16,7 @@ internal final class TestAllTests: XCTestCase {
 
     func test_all_passes_when_it_should_2() {
         var count = 0
-        TestSuite().testAll(genTrue, genTrue) { assert, v1, v2 in
+        testAll(genTrue, genTrue) { assert, v1, v2 in
             count += 1
             assert(v1 && v2)
         }
@@ -26,7 +26,7 @@ internal final class TestAllTests: XCTestCase {
 
     func test_all_passes_when_it_should_3() {
         var count = 0
-        TestSuite().testAll(genTrue, genTrue, genTrue) { assert, v1, v2, v3 in
+        testAll(genTrue, genTrue, genTrue) { assert, v1, v2, v3 in
             count += 1
             assert(v1 && v2 && v3)
         }
@@ -36,7 +36,7 @@ internal final class TestAllTests: XCTestCase {
 
     func test_all_passes_when_it_should_4() {
         var count = 0
-        TestSuite().testAll(genTrue, genTrue, genTrue, genTrue) { assert, v1, v2, v3, v4 in
+        testAll(genTrue, genTrue, genTrue, genTrue) { assert, v1, v2, v3, v4 in
             count += 1
             assert(v1 && v2 && v3 && v4)
         }
@@ -46,7 +46,7 @@ internal final class TestAllTests: XCTestCase {
 
     func test_all_passes_when_it_should_5() {
         var count = 0
-        TestSuite().testAll(genTrue, genTrue, genTrue, genTrue, genTrue) { assert, v1, v2, v3, v4, v5 in
+        testAll(genTrue, genTrue, genTrue, genTrue, genTrue) { assert, v1, v2, v3, v4, v5 in
             count += 1
             assert(v1 && v2 && v3 && v4 && v5)
         }
@@ -56,7 +56,7 @@ internal final class TestAllTests: XCTestCase {
 
     func test_all_passes_when_it_should_6() {
         var count = 0
-        TestSuite().testAll(genTrue, genTrue, genTrue, genTrue, genTrue, genTrue) { assert, v1, v2, v3, v4, v5, v6 in
+        testAll(genTrue, genTrue, genTrue, genTrue, genTrue, genTrue) { assert, v1, v2, v3, v4, v5, v6 in
             count += 1
             assert(v1 && v2 && v3 && v4 && v5 && v6)
         }
@@ -66,19 +66,17 @@ internal final class TestAllTests: XCTestCase {
 
     func test_all_passes_when_it_should_7() {
         var count = 0
-        TestSuite()
-            .testAll(genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue) { assert, v1, v2, v3, v4, v5, v6, v7 in
-                count += 1
-                assert(v1 && v2 && v3 && v4 && v5 && v6 && v7)
-            }
+        testAll(genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue) { assert, v1, v2, v3, v4, v5, v6, v7 in
+            count += 1
+            assert(v1 && v2 && v3 && v4 && v5 && v6 && v7)
+        }
 
         XCTAssertEqual(TestConfig.default().maxIterations, count)
     }
 
     func test_all_passes_when_it_should_8() {
         var count = 0
-        TestSuite()
-            .testAll(genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue)
+        testAll(genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue)
             { assert, v1, v2, v3, v4, v5, v6, v7, v8 in
                 count += 1
                 assert(v1 && v2 && v3 && v4 && v5 && v6 && v7 && v8)
@@ -89,8 +87,7 @@ internal final class TestAllTests: XCTestCase {
 
     func test_all_passes_when_it_should_9() {
         var count = 0
-        TestSuite()
-            .testAll(genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue)
+        testAll(genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue)
             { assert, v1, v2, v3, v4, v5, v6, v7, v8, v9 in
                 count += 1
                 assert(v1 && v2 && v3 && v4 && v5 && v6 && v7 && v8 && v9)
@@ -101,8 +98,7 @@ internal final class TestAllTests: XCTestCase {
 
     func test_all_passes_when_it_should_10() {
         var count = 0
-        TestSuite()
-            .testAll(genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue)
+        testAll(genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue, genTrue)
             { assert, v1, v2, v3, v4, v5, v6, v7, v8, v9, v10 in
                 count += 1
                 assert(v1 && v2 && v3 && v4 && v5 && v6 && v7 && v8 && v9 && v10)
@@ -113,7 +109,7 @@ internal final class TestAllTests: XCTestCase {
 
     func test_all_fails_when_it_should() {
         XCTExpectFailure()
-        TestSuite().testAll((1 ... 100).arbitrary) { assert, a in
+        testAll((1 ... 100).arbitrary) { assert, a in
             assert(a < 1)
         }
     }
