+++ Tests/GenythingTests/Generators/Generators_CollectTests.swift
@@ -4,7 +4,7 @@ import XCTest
 
 final class Generators_CollectTests: XCTestCase {
     func test_collecting_to_alternate_generators() {
-        TestSuite().testAllSatisfy(Generators.collect(
+        testAllSatisfy(Generators.collect(
             Generators.constant(0),
             Generators.constant(1)
         )) { values in
