+++ Tests/GenythingTests/Generators/Generators_IterateTests.swift
@@ -5,7 +5,7 @@ import XCTest
 final class Generators_IterateTests: XCTestCase {
     func test_the_sequence_generates_expected_values() {
         var expected = 0
-        TestSuite().testAllSatisfy(Generators.iterate(0 ... Int.max)) {
+        testAllSatisfy(Generators.iterate(0 ... Int.max)) {
             defer { expected += 1 }
             return expected == $0
         }
@@ -14,7 +14,7 @@ final class Generators_IterateTests: XCTestCase {
     func test_exhausted_sequence_generates_nil() {
         let countNonNil = 10
         var i = 0
-        TestSuite().testAllSatisfy(Generators.iterate(0 ..< 10)) {
+        testAllSatisfy(Generators.iterate(0 ..< 10)) {
             defer { i += 1 }
             if i < countNonNil {
                 return $0 != nil
@@ -24,7 +24,7 @@ final class Generators_IterateTests: XCTestCase {
     }
 
     func test_empty_sequence_generates_nil() {
-        TestSuite().testAllSatisfy(Generators.iterate([])) {
+        testAllSatisfy(Generators.iterate([])) {
             $0 == nil
         }
     }
