+++ Tests/GenythingTests/Operators/Generator_ExpandTests.swift
@@ -6,7 +6,7 @@ final class Generator_ExpandTests: XCTestCase {
     func test_expand_toSize_createsWithCorrectSize() {
         let size = 10
         let gen = String.arbitrary.expand(toSize: size)
-        TestSuite().testAllSatisfy(gen) { s in
+        testAllSatisfy(gen) { s in
             s.count == size
         }
     }
@@ -14,7 +14,7 @@ final class Generator_ExpandTests: XCTestCase {
     func test_expand_toSizeInRange_createsInsideCorrectBounds() {
         let range: Range<Int> = 0 ..< 5
         let gen = String.arbitrary.expand(toSizeInRange: range)
-        TestSuite().testAllSatisfy(gen) { s in
+        testAllSatisfy(gen) { s in
             range.lowerBound <= s.count && s.count <= range.upperBound
         }
     }
@@ -22,7 +22,7 @@ final class Generator_ExpandTests: XCTestCase {
     func test_expand_toSizeInClosedRange_createsInsideCorrectBounds() {
         let range: ClosedRange<Int> = 0 ... 5
         let gen = String.arbitrary.expand(toSizeInRange: range)
-        TestSuite().testAllSatisfy(gen) { s in
+        testAllSatisfy(gen) { s in
             range.lowerBound <= s.count && s.count <= range.upperBound
         }
     }
