+++ Tests/GenythingTests/Operators/Generator_FilterTests.swift
@@ -7,7 +7,7 @@ final class Generator_FilterTests: XCTestCase {
         let isEven = { ($0 % 2) == 0 } // `1/2` chance of occuring
         let gen = Int.arbitrary.filter(isEven)
 
-        TestSuite().testAllSatisfy(gen) {
+        testAllSatisfy(gen) {
             isEven($0)
         }
     }
@@ -21,7 +21,7 @@ final class Generator_FilterTests: XCTestCase {
             let isDigit = { $0 >= 0 && $0 < 10 } // `10/1.8446744e+19` chance of occuring
             let gen = Int.arbitrary.filter(isDigit)
 
-            TestSuite().testAllSatisfy(gen) {
+            testAllSatisfy(gen) {
                 isDigit($0)
             }
