+++ Tests/TrickeryTests/FakeCharactersTests.swift
@@ -4,19 +4,19 @@ import XCTest
 
 class FakeCharactersTests: XCTestCase {
     func test_lowercase() {
-        TestSuite().testAllSatisfy(Fake.Characters.lowercase) {
+        testAllSatisfy(Fake.Characters.lowercase) {
             $0.isLowercase
         }
     }
 
     func test_uppercase() {
-        TestSuite().testAllSatisfy(Fake.Characters.uppercase) {
+        testAllSatisfy(Fake.Characters.uppercase) {
             $0.isUppercase
         }
     }
 
     func test_alphanumeric() {
-        TestSuite().testAllSatisfy(Fake.Characters.alphanumeric) {
+        testAllSatisfy(Fake.Characters.alphanumeric) {
             $0.isNumber || $0.isLetter
         }
     }
