+++ Tests/TrickeryTests/FakeIDTests.swift
@@ -4,7 +4,7 @@ import XCTest
 
 class FakeIDTests: XCTestCase {
     func test_length() {
-        TestSuite().testAllSatisfy(Fake.ID.uuid) {
+        testAllSatisfy(Fake.ID.uuid) {
             $0.uuidString.count == 36
         }
     }
