+++ Tests/TrickeryTests/FakeLocationTests.swift
@@ -5,21 +5,21 @@ import XCTest
 
 class FakeLocationTests: XCTestCase {
     func test_latitude() {
-        TestSuite().testAll(Fake.Locations.latitude) { assert, val in
+        testAll(Fake.Locations.latitude) { assert, val in
             assert(val >= -90.0)
             assert(val <= 90.0)
         }
     }
 
     func test_longitude() {
-        TestSuite().testAll(Fake.Locations.longitude) { assert, val in
+        testAll(Fake.Locations.longitude) { assert, val in
             assert(val >= -180.0)
             assert(val <= 180.0)
         }
     }
 
     func test_coordinate() {
-        TestSuite().testAllSatisfy(Fake.Locations.coordinate) {
+        testAllSatisfy(Fake.Locations.coordinate) {
             CLLocationCoordinate2DIsValid($0)
         }
     }
