+++ Tests/TrickeryTests/FakePersonNamesTest.swift
@@ -4,7 +4,7 @@ import XCTest
 
 class FakeNamesTests: XCTestCase {
     func test_names_full() {
-        TestSuite().testAll(Fake.PersonNames.full()) { assert, value in
+        testAll(Fake.PersonNames.full()) { assert, value in
             let names = value.components(separatedBy: .whitespaces)
             names.forEach {
                 assert.equal(true, $0.first?.isUppercase)
@@ -13,7 +13,7 @@ class FakeNamesTests: XCTestCase {
     }
 
     func test_names_full_number_of_words() {
-        TestSuite().testAllSatisfy(Fake.PersonNames.full()) {
+        testAllSatisfy(Fake.PersonNames.full()) {
             // Some names like "Jo Ann" can have 2 words
             let numberOfWords = $0.components(separatedBy: .whitespaces).count
